Implement #define expansion
This adds a stack of scanner states to SPScanner. Everytime an
identifier is found to be defined in the preprocessor, let the scanner
scan from the replacement string. If the replacement ended (EOF) resume
with lexing the original text which used the define.

The preprocessor only knows about String and Integer types. No need to
lex other types (float, char).